HEAT-30 feat : translation controller 구현

+++ src/main/java/com/hyundaiautoever/HEAT/v1/dto/TranslationDto.java
@@ -8,7 +8,7 @@ import lombok.Getter;
 @Builder
 public class TranslationDto {
 
-    private Integer translationNo;
+    private Long translationNo;
     private String userId;
     private String requestLanguageName;
     private String resultLanguageName;
